Backend-Sqlite/requirements.txt
# requirements.txt
# pip install -r requirements.txt
# uvicorn main:app --reload
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib